move some more CPU related work to CPU class previously found in TsrConstructor in preparation for allowing CPU data to be created from Data interface
