+++ src/main/java/neureka/dtype/DataType.java
@@ -148,6 +148,10 @@ public final class DataType<T>
     public Class<?> dataArrayType() {
         if ( this.typeClassImplements( NumericType.class ) )
             return ( (NumericType<?,?,?,?>) Objects.requireNonNull( getTypeClassInstance(NumericType.class) ) ).holderArrayType();
+        else if ( this.getItemTypeClass() == Character.class )
+            return char[].class;
+        else if ( this.getItemTypeClass() == Boolean.class )
+            return boolean[].class;
         else
             return Object[].class;
     }
